+++ src/functions/discorddocs.ts
@@ -7,8 +7,14 @@ import { fetch } from 'undici';
 import { Response } from 'polka';
 import { stringify } from 'querystring';
 import { AlgoliaSearchResult } from '../types/algolia';
-import { API_BASE_ALGOLIA, DEFAULT_ALGOLIA_RESULT_AMOUNT, EMOJI_ID_CLYDE_BLURPLE } from '../util/constants';
-import { prepareErrorResponse, prepareResponse } from '../util/respond';
+
+import {
+	prepareErrorResponse,
+	prepareResponse,
+	API_BASE_ALGOLIA,
+	DEFAULT_ALGOLIA_RESULT_AMOUNT,
+	EMOJI_ID_CLYDE_BLURPLE,
+} from '../util';
 
 export async function discordDeveloperDocs(
 	response: Response,
