+++ src/functions/node.ts
@@ -16,9 +16,7 @@ import { fetch } from 'undici';
 import { Response } from 'polka';
 import TurndownService from 'turndown';
 import type { NodeDocs } from '../types/NodeDocs';
-import { API_BASE_NODE, EMOJI_ID_NODE } from '../util/constants';
-import { logger } from '../util/logger';
-import { prepareErrorResponse, prepareResponse } from '../util/respond';
+import { API_BASE_NODE, EMOJI_ID_NODE, logger, prepareErrorResponse, prepareResponse } from '../util';
 
 const td = new TurndownService({ codeBlockStyle: 'fenced' });
