+++ src/functions/tag.ts
@@ -1,14 +1,18 @@
-import { prepareErrorResponse, prepareResponse } from '../util/respond';
 import { Response } from 'polka';
 import Collection from '@discordjs/collection';
 import { distance } from 'fastest-levenshtein';
-import { PREFIX_SUCCESS, REMOTE_TAG_URL } from '../util/constants';
+import {
+	PREFIX_SUCCESS,
+	REMOTE_TAG_URL,
+	prepareErrorResponse,
+	prepareResponse,
+	logger,
+	suggestionString,
+} from '../util';
 import { readFileSync } from 'fs';
 import { join } from 'path';
 import * as TOML from '@ltd/j-toml';
-import { logger } from '../util/logger';
 import { fetch } from 'undici';
-import { suggestionString } from '../util';
 
 export interface Tag {
 	keywords: string[];
