+++ src/index.ts
@@ -1,17 +1,25 @@
-import polka, { Middleware } from 'polka';
-import { logger } from './util/logger';
-import { jsonParser } from './util/jsonParser';
-import { prepareAck, prepareErrorResponse, prepareResponse } from './util/respond';
+import polka, { Middleware, NextHandler, Request, Response } from 'polka';
+import {
+	logger,
+	jsonParser,
+	prepareAck,
+	prepareErrorResponse,
+	prepareResponse,
+	API_BASE_MDN,
+	ArgumentsOf,
+	EMOJI_ID_CLYDE_BLURPLE,
+	EMOJI_ID_GUIDE,
+	PREFIX_BUG,
+	PREFIX_TEAPOT,
+	transformInteraction,
+} from './util';
 import { djsDocs } from './functions/docs';
 import { mdnSearch } from './functions/mdn';
 import { nodeSearch } from './functions/node';
-import { API_BASE_MDN, EMOJI_ID_CLYDE_BLURPLE, EMOJI_ID_GUIDE, PREFIX_BUG, PREFIX_TEAPOT } from './util/constants';
 import { loadTags, reloadTags, showTag, Tag } from './functions/tag';
 import Collection from '@discordjs/collection';
 import { Doc } from 'discordjs-docs-parser';
 import { InteractionType, APIInteraction, ApplicationCommandType } from 'discord-api-types/v10';
-import { transformInteraction } from './util/interactionOptions';
-import { ArgumentsOf } from './util/argumentsOf';
 import { DiscordDocsCommand } from './interactions/discorddocs';
 import { GuideCommand } from './interactions/guide';
 import { MdnCommand } from './interactions/mdn';
@@ -23,7 +31,63 @@ import { resolveOptionsToDocsAutoComplete, djsDocsAutoComplete } from './functio
 import { mdnAutoComplete } from './functions/autocomplete/mdnAutoComplete';
 import { algoliaAutoComplete } from './functions/autocomplete/algoliaAutoComplete';
 import { algoliaResponse } from './functions/algoliaResponse';
-import { verify } from './util/verify';
+import { webcrypto } from 'node:crypto';
+
+// @ts-expect-error
+const { subtle } = webcrypto;
+
+const encoder = new TextEncoder();
+
+function hex2bin(hex: string) {
+	const buf = new Uint8Array(Math.ceil(hex.length / 2));
+	for (let i = 0; i < buf.length; i++) {
+		buf[i] = parseInt(hex.slice(i * 2, i * 2 + 2), 16);
+	}
+	return buf;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+const PUBKEY = subtle.importKey(
+	'raw',
+	hex2bin(process.env.DISCORD_PUBKEY!),
+	{
+		name: 'NODE-ED25519',
+		namedCurve: 'NODE-ED25519',
+		public: true,
+	},
+	true,
+	['verify'],
+);
+
+async function verify(req: Request, res: Response, next: NextHandler) {
+	if (!req.headers['x-signature-ed25519']) {
+		res.writeHead(401);
+		return res.end();
+	}
+	const signature = req.headers['x-signature-ed25519'] as string;
+	const timestamp = req.headers['x-signature-timestamp'] as string;
+
+	if (!signature || !timestamp) {
+		res.writeHead(401);
+		return res.end();
+	}
+
+	const hexSignature = hex2bin(signature);
+
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+	const isValid = await subtle.verify(
+		'NODE-ED25519',
+		await PUBKEY,
+		hexSignature,
+		encoder.encode(timestamp + req.rawBody),
+	);
+
+	if (!isValid) {
+		res.statusCode = 401;
+		return res.end();
+	}
+	void next();
+}
 
 type CommandName = 'discorddocs' | 'docs' | 'guide' | 'invite' | 'mdn' | 'node' | 'tag' | 'tagreload';
 type CommandAutoCompleteName = 'docs' | 'tag' | 'mdn' | 'guide' | 'discorddocs';
