+++ src/util/truncate.ts
@@ -1,0 +1,14 @@
+export function truncate(text: string, len: number, splitChar = ' '): string {
+	if (text.length <= len) return text;
+	const words = text.split(splitChar);
+	const res: string[] = [];
+	for (const word of words) {
+		const full = res.join(splitChar);
+		if (full.length + word.length + 1 <= len - 3) {
+			res.push(word);
+		}
+	}
+
+	const resText = res.join(splitChar);
+	return resText.length === text.length ? resText : `${resText.trim()}...`;
+}
